+++ core-ng/src/main/java/core/framework/api/validate/ValidationException.java
@@ -1,5 +1,6 @@
 package core.framework.api.validate;
 
+import core.framework.api.log.ErrorCode;
 import core.framework.api.log.Warning;
 import core.framework.api.util.Strings;
 
@@ -9,12 +10,17 @@ import java.util.Map;
  * @author neo
  */
 @Warning
-public class ValidationException extends RuntimeException {
+public class ValidationException extends RuntimeException implements ErrorCode {
     private static final long serialVersionUID = 9215299700445046388L;
     public final Map<String, String> errors;
 
     public ValidationException(Map<String, String> errors) {
         super(Strings.format("validation failed, error={}", errors));
         this.errors = errors;
+    }
+
+    @Override
+    public String errorCode() {
+        return "VALIDATION_ERROR";
     }
 }