+++ core-ng/src/main/java/core/framework/impl/web/HTTPServerErrorHandler.java
@@ -14,7 +14,6 @@ import core.framework.api.web.exception.ConflictException;
 import core.framework.api.web.exception.ForbiddenException;
 import core.framework.api.web.exception.MethodNotAllowedException;
 import core.framework.api.web.exception.NotFoundException;
-import core.framework.api.web.exception.RemoteServiceException;
 import core.framework.api.web.exception.UnauthorizedException;
 import core.framework.impl.web.exception.ErrorResponse;
 import core.framework.impl.web.request.RequestImpl;
@@ -110,8 +109,6 @@ public class HTTPServerErrorHandler {
         response.message = e.getMessage();
         response.stackTrace = Exceptions.stackTrace(e);
         if (e instanceof ErrorCode) response.errorCode = ((ErrorCode) e).errorCode();
-        else if (e instanceof ValidationException) response.errorCode = "VALIDATION_ERROR";
-        else if (e instanceof RemoteServiceException) response.errorCode = ((RemoteServiceException) e).errorCode;
         else response.errorCode = "INTERNAL_ERROR";
         return response;
     }